# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CDT Jamaica Management API with Odoo Community integration
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /classes:
    get:
      operationId: listClasses
      tags: [classes]
      summary: List all classes
      parameters:
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: List of classes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DanceClass"
    post:
      operationId: createClass
      tags: [classes]
      summary: Legacy class creation endpoint; linked records are managed in Odoo
      deprecated: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassInput"
      responses:
        "201":
          description: Created class
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DanceClass"
        "409":
          description: Class and enrollment changes must be made in Odoo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /classes/{id}:
    get:
      operationId: getClass
      tags: [classes]
      summary: Get a class by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Class
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DanceClass"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateClass
      tags: [classes]
      summary: Legacy class update endpoint; linked records are managed in Odoo
      deprecated: true
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassUpdate"
      responses:
        "200":
          description: Updated class
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DanceClass"
        "409":
          description: Class and enrollment changes must be made in Odoo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteClass
      tags: [classes]
      summary: Legacy class deletion endpoint; linked records are managed in Odoo
      deprecated: true
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "409":
          description: Class and enrollment changes must be made in Odoo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Contacts (Odoo) ─────────────────────────────────────────────────────
  /contacts:
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List contacts from Odoo
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: type
          schema:
            type: ["string", "null"]
            enum: [individual, company, null]
      responses:
        "200":
          description: List of contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
    post:
      operationId: createContact
      tags: [contacts]
      summary: Create a contact in Odoo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Created contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
  /contacts/{id}:
    get:
      operationId: getContact
      tags: [contacts]
      summary: Get a contact by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateContact
      tags: [contacts]
      summary: Update a contact in Odoo
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactUpdate"
      responses:
        "200":
          description: Updated contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "400":
          description: Invalid contact update
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Attempted to change a public profile field still managed by Sanity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Donations (Odoo) ────────────────────────────────────────────────────
...
  /content/team:
    get:
      operationId: listTeamMembers
      tags: [content]
      summary: List website team members mirrored into Odoo
      parameters:
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: Team members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamMember"
  /content/repertoire:
    get:
      operationId: listRepertoireItems
      tags: [content]
      summary: List repertoire mirrored into Odoo
      parameters:
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: Repertoire items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RepertoireItem"
  /content/media:
    get:
      operationId: listMediaItems
      tags: [content]
      summary: List website media mirrored into Odoo
      parameters:
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: Media items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MediaItem"
  /content/website-settings:
    get:
      operationId: getWebsiteSettings
      tags: [content]
      summary: Get mirrored website settings and sync status
      responses:
        "200":
          description: Website settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebsiteSetting"
        "404":
          description: Website settings not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    DanceClass:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        style:
          type: string
        instructor:
          type: string
        schedule:
          type: string
        location:
          type: string
        capacity:
          type: integer
        enrolledCount:
          type: integer
        status:
          type: string
          enum: [active, inactive]
        description:
          type: ["string", "null"]
        ageGroup:
          type: ["string", "null"]
        fee:
          type: ["number", "null"]
        createdAt:
          type: string
        classStatus:
          type: string
        academicYear:
          type: ["string", "null"]
        academicTerm:
          type: ["string", "null"]
        deliveryMode:
          type: string
        studentIds:
          type: array
          items:
            type: integer
        studentNames:
          type: array
          items:
            type: string
        currency:
          type: string
        sampleData:
          type: boolean
      required: [id, name, style, instructor, schedule, location, capacity, enrolledCount, status, createdAt, classStatus, academicYear, academicTerm, deliveryMode, studentIds, studentNames, currency, sampleData]
    ClassInput:
      type: object
      properties:
        name:
          type: string
        style:
          type: string
        instructor:
          type: string
        schedule:
          type: string
        location:
          type: string
        capacity:
          type: integer
        status:
          type: string
          enum: [active, inactive]
        description:
          type: string
        ageGroup:
          type: string
        fee:
          type: number
      required: [name, style, instructor, schedule, location, capacity]
    ClassUpdate:
      type: object
      properties:
        name:
          type: string
        style:
          type: string
        instructor:
          type: string
        schedule:
          type: string
        location:
          type: string
        capacity:
          type: integer
        status:
          type: string
          enum: [active, inactive]
        description:
          type: string
        ageGroup:
          type: string
        fee:
          type: number

    # ── Contact ────────────────────────────────────────────────────────────
    Contact:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        type:
          type: string
          enum: [individual, company]
        company:
          type: ["string", "null"]
        street:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        odooId:
          type: ["string", "null"]
        createdAt:
          type: string
        managedBySanity:
          type: boolean
          description: Whether public profile identity fields are still managed by the Sanity compatibility sync.
      required: [id, name, type, createdAt, managedBySanity]
    ContactInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
        phone:
          type: string
        type:
          type: string
          enum: [individual, company]
        company:
          type: string
        street:
          type: string
        city:
          type: string
        country:
          type: string
      required: [name]
    ContactUpdate:
      type: object
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
        phone:
          type: string
        type:
          type: string
          enum: [individual, company]
        company:
          type: string
        street:
          type: string
        city:
          type: string
        country:
          type: string

    # ── Donation ───────────────────────────────────────────────────────────
...
    TeamMember:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        memberType:
          type: string
          enum: [dancer, management, board]
        role:
          type: ["string", "null"]
        yearsActive:
          type: ["string", "null"]
        term:
          type: ["string", "null"]
        featured:
          type: boolean
        sequence:
          type: integer
        headshotUrl:
          type: ["string", "null"]
        sourceUpdatedAt:
          type: ["string", "null"]
        managedBySanity:
          type: boolean
      required: [id, name, memberType, featured, sequence, managedBySanity]
    RepertoireItem:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        subtitle:
          type: ["string", "null"]
        slug:
          type: string
        year:
          type: integer
        runtime:
          type: ["string", "null"]
        choreographer:
          type: ["string", "null"]
        genre:
          type: ["string", "null"]
        stylePeriod:
          type: ["string", "null"]
        youtubeId:
          type: ["string", "null"]
        thumbnailUrl:
          type: ["string", "null"]
        heroImageUrl:
          type: ["string", "null"]
        sourceUpdatedAt:
          type: ["string", "null"]
        managedBySanity:
          type: boolean
      required: [id, title, slug, year, managedBySanity]
    MediaItem:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: ["string", "null"]
        videoType:
          type: ["string", "null"]
        duration:
          type: ["string", "null"]
        publishedAt:
          type: ["string", "null"]
        featured:
          type: boolean
        tags:
          type: array
          items:
            type: string
        category:
          type: ["string", "null"]
        videoUrl:
          type: ["string", "null"]
        vimeoUrl:
          type: ["string", "null"]
        youtubeUrl:
          type: ["string", "null"]
        thumbnailUrl:
          type: ["string", "null"]
        sourceUpdatedAt:
          type: ["string", "null"]
        managedBySanity:
          type: boolean
      required: [id, title, featured, tags, managedBySanity]
    WebsiteSetting:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        lightLogoUrl:
          type: ["string", "null"]
        darkLogoUrl:
          type: ["string", "null"]
        heroImageUrl:
          type: ["string", "null"]
        homeHeroImageUrl:
          type: ["string", "null"]
        lastSyncAt:
          type: ["string", "null"]
        lastSyncStatus:
          type: ["string", "null"]
          enum: [success, failed, null]
        lastSyncMessage:
          type: ["string", "null"]
        sourceUpdatedAt:
          type: ["string", "null"]
        managedBySanity:
          type: boolean
      required: [id, title, managedBySanity]

    # ── Performance ────────────────────────────────────────────────────────
...